Toward replacing show[Code] with inferToken
