+++ tests/src/test/scala/tokens/InferPartialSuite.scala
@@ -10,6 +10,7 @@ class InferPartialSuite extends FunSuite {
       """import dummy.test
         |/* This high-level comment should be persisted */
         |case class Test(string: String)
+        |// TODO: this is a second high-level comment.
         |object Test {
         |  val string = "This is a string"  
         |  def funToReplace(x: Int, y: String) {
@@ -25,24 +26,26 @@ class InferPartialSuite extends FunSuite {
 
     // Forcing synthetic tokens using TQL
     val trans = (transform {
-      case t: impl.Defn.Def if t.name.value == "funToReplace" => 
+      case t: impl.Defn.Def if t.name.value == "funToReplace" =>
         t.copy(name = impl.Term.Name("aNewName")) andCollect Unit
       case t: impl.Defn.Val =>
         t.copy() andCollect Unit
+      case t: impl.Defn.Object if t.name.value == "Test" =>
+        t.copy(name = impl.Term.Name("Test2")) andCollect Unit
     }).topDown
 
     val transformed = trans(testTree)
     val newCode = transformed.tree.get.tokens.map(_.show[Code]).mkString
-    println("Printing original tokens:")
-    println(testTree.tokens.map(_.code).mkString)
-    println("-----------")
-    println("Printing tokens corresponding to stats in Source:")
-    testTree.asInstanceOf[impl.Source].stats.foreach {s => println(s.tokens.map(_.code).mkString + "\n")}
-    println("-----------")
-    println("Printing tokens after modification by TQL:")
-    println(newCode)
-    assert(newCode.contains("/* Just creating a new case object */"))
+    
+    /* Cheking that highl-level comments are persisted */
     assert(newCode.contains("/* This high-level comment should be persisted */"))
+    assert(newCode.contains("// TODO: this is a second high-level comment."))
+
+    /* Checking that inner comment of unmodified parts are persisted */
+    assert(newCode.contains("/* Just creating a new case object */"))
+
+    /* Checking that indentation is persisted for unmodified part */
+    assert(newCode.contains("                                  "))
   }
 
 }
